requirements.txt
Flask
Werkzeug
python-dotenv
flask-wtf
flask-limiter
flask-migrate